# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CareOS API — home care operating system foundation.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      parameters:
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/ClientStatus"
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientDetail"
    patch:
      operationId: updateClient
      tags: [clients]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}/authorizations:
    get:
      operationId: listClientAuthorizations
      tags: [authorizations]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Authorization"
    post:
      operationId: createClientAuthorization
      tags: [authorizations]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAuthorizationBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Authorization"
...
  /labor-rules/active:
    get:
      operationId: getActiveLaborRule
      tags: [rules]
      summary: Currently active labor rule for the agency
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LaborRuleSet"
    post:
      operationId: setActiveLaborRule
      tags: [rules]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetActiveLaborRuleBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LaborRuleSet"
...
components:
  schemas:
...
    ClientStatus:
      type: string
      enum: [PROSPECT, ACTIVE, ON_HOLD, DISCHARGED]
    PayerType:
      type: string
      enum: [PRIVATE_PAY, VA_CCN, MEDICAID_HCBS, COUNTY_IHSS, LTC_INSURANCE]
    Client:
      type: object
      properties:
        id: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        dob: { type: string, format: date }
        phone: { type: string, nullable: true }
        email: { type: string, nullable: true }
        addressLine1: { type: string, nullable: true }
        city: { type: string, nullable: true }
        state: { type: string, nullable: true }
        postalCode: { type: string, nullable: true }
        primaryPayer:
          $ref: "#/components/schemas/PayerType"
        status:
          $ref: "#/components/schemas/ClientStatus"
        intakeDate: { type: string, format: date, nullable: true }
        languages:
          type: array
          items: { type: string }
        carePreferences: { type: string, nullable: true }
        allergies: { type: string, nullable: true }
        emergencyContactName: { type: string, nullable: true }
        emergencyContactPhone: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
      required:
        [id, firstName, lastName, dob, primaryPayer, status, languages, createdAt]
    ClientDetail:
      allOf:
        - $ref: "#/components/schemas/Client"
        - type: object
          properties:
            authorizations:
              type: array
              items:
                $ref: "#/components/schemas/Authorization"
            recentVisits:
              type: array
              items:
                $ref: "#/components/schemas/Visit"
          required: [authorizations, recentVisits]
    CreateClientBody:
      type: object
      properties:
        firstName: { type: string }
        lastName: { type: string }
        dob: { type: string, format: date }
        phone: { type: string }
        email: { type: string }
        addressLine1: { type: string }
        city: { type: string }
        state: { type: string }
        postalCode: { type: string }
        primaryPayer:
          $ref: "#/components/schemas/PayerType"
        languages:
          type: array
          items: { type: string }
        carePreferences: { type: string }
        allergies: { type: string }
        emergencyContactName: { type: string }
        emergencyContactPhone: { type: string }
      required: [firstName, lastName, dob, primaryPayer]
    UpdateClientBody:
      type: object
      properties:
        firstName: { type: string }
        lastName: { type: string }
        phone: { type: string }
        email: { type: string }
        addressLine1: { type: string }
        city: { type: string }
        state: { type: string }
        postalCode: { type: string }
        status:
          $ref: "#/components/schemas/ClientStatus"
        primaryPayer:
          $ref: "#/components/schemas/PayerType"
        carePreferences: { type: string }
        allergies: { type: string }
        emergencyContactName: { type: string }
        emergencyContactPhone: { type: string }
    AuthorizationStatus:
      type: string
      enum: [ACTIVE, EXPIRING_SOON, EXPIRED, EXHAUSTED]
    Authorization:
      type: object
      properties:
        id: { type: string }
        clientId: { type: string }
        clientName: { type: string }
        payer:
          $ref: "#/components/schemas/PayerType"
        authNumber: { type: string }
        issuedDate: { type: string, format: date }
        expirationDate: { type: string, format: date }
        approvedHoursPerWeek: { type: number }
        approvedHoursTotal: { type: number }
        hoursUsed: { type: number }
        hoursRemaining: { type: number }
        scopeOfCare:
          type: array
          items: { type: string }
        status:
          $ref: "#/components/schemas/AuthorizationStatus"
        daysUntilExpiration: { type: integer }
      required:
        [
          id,
          clientId,
          clientName,
          payer,
          authNumber,
          issuedDate,
          expirationDate,
          approvedHoursPerWeek,
          approvedHoursTotal,
          hoursUsed,
          hoursRemaining,
          scopeOfCare,
          status,
          daysUntilExpiration,
        ]
    CreateAuthorizationBody:
      type: object
      properties:
        payer:
          $ref: "#/components/schemas/PayerType"
        authNumber: { type: string }
        issuedDate: { type: string, format: date }
        expirationDate: { type: string, format: date }
        approvedHoursPerWeek: { type: number }
        approvedHoursTotal: { type: number }
        scopeOfCare:
          type: array
          items: { type: string }
      required:
        [
          payer,
          authNumber,
          issuedDate,
          expirationDate,
          approvedHoursPerWeek,
          approvedHoursTotal,
        ]
...
    VisitVerificationStatus:
      type: string
      enum: [PENDING, VERIFIED, EXCEPTION, REJECTED]
    ClockMethod:
      type: string
      enum: [GPS, TELEPHONY, FOB, MANUAL]
    Visit:
      type: object
      properties:
        id: { type: string }
        scheduleId: { type: string, nullable: true }
        caregiverId: { type: string }
        caregiverName: { type: string }
        clientId: { type: string }
        clientName: { type: string }
        clockInTime: { type: string, format: date-time, nullable: true }
        clockInLat: { type: number, nullable: true }
        clockInLng: { type: number, nullable: true }
        clockInMethod:
          $ref: "#/components/schemas/ClockMethod"
        clockOutTime: { type: string, format: date-time, nullable: true }
        clockOutLat: { type: number, nullable: true }
        clockOutLng: { type: number, nullable: true }
        clockOutMethod:
          $ref: "#/components/schemas/ClockMethod"
        durationMinutes: { type: integer, nullable: true }
        tasksCompleted:
          type: array
          items: { type: string }
        caregiverNotes: { type: string, nullable: true }
        supervisorNotes: { type: string, nullable: true }
        verificationStatus:
          $ref: "#/components/schemas/VisitVerificationStatus"
        exceptionReason: { type: string, nullable: true }
        geoFenceMatch: { type: boolean }
      required:
        [
          id,
          caregiverId,
          caregiverName,
          clientId,
          clientName,
          clockInMethod,
          clockOutMethod,
          tasksCompleted,
          verificationStatus,
          geoFenceMatch,
        ]
...
    LaborRuleSet:
      type: object
      properties:
        id: { type: string }
        state: { type: string }
        name: { type: string }
        description: { type: string }
        version: { type: string }
        overtimeThresholdDailyMinutes: { type: integer, nullable: true }
        overtimeThresholdWeeklyMinutes: { type: integer, nullable: true }
        doubleTimeThresholdDailyMinutes: { type: integer, nullable: true }
        seventhDayConsecutiveRule: { type: boolean }
        travelTimeBillable: { type: boolean }
        isActive: { type: boolean }
      required:
        [
          id,
          state,
          name,
          description,
          version,
          seventhDayConsecutiveRule,
          travelTimeBillable,
          isActive,
        ]
    SetActiveLaborRuleBody:
      type: object
      properties:
        ruleId: { type: string }
      required: [ruleId]
...